grid-control: retire path-changed event -> the reactive var current_path should be used instead (see reason in description)
By using the reactive var we avoid the need to take care of situations
where due to grid flush active path is temporarily reseted.
emit is synchronous -> all changes will be catched by it
changes to the reactive var on the other hand will occur only in the
next Tracker.flush - which for sure will be no sooner than the the next
tick.