grid-data: behavior update: stop bypassing flush process for simple items fields updates (see commit description)
From now on, all changes should go through the flush process to make
sure the flush lock mechanism is honored.

We need to be able to rely on the fact that no grid updates are occuring
during flush lock. Without this we have no way to tell when a certain
field is getting invalidated and some depending structure can't be
destroyed. (open owners dropdown is an example for depending structure)

The comment in the removed code is irrelevant since today even filter
state related grid updates go through the flush mechanizem. If anything
it's another reason to change this behavior which today can even look
buggy (Example: filtered field change while lock is on - the field
updated but the row removed from the grid only upon flush lock release).