[bugfix] grid-control: fix a drag-n-drop edge case
jQueryUI keeps information about the item previous to the original
position
of the moved item to be able to reposition it in its original place in
case
of cancel.
Expanding a path might change the previous item if the original position
of the
moved item was just before the expanded sub-tree.
Therefore we need to update the internal data structure of jquery ui
sortable
to make sure we won't run into bugs.