grid-data-com: all grid-data-com Meteor methods been refactored as the GridDataCom object API methods to allow accessing their logic directly
* All API methods get as their last param a new argument "perform_as"
which replaces the @userId of the Meteor Method `this` context.
    * With that packages that uses the grid-data-com API can easily
configure this setting.
* Middlewares API changed:
    * from now on the `this` variable assigned to grid-data-com
middleware will be the grid-data-com object and the Meteor method's this
(as now, there might not be Meteor method that trigger the action at
all).
    * Additional argument was added to each one of the middlewares hooks
that pass the user_id of the user performing the action (the
perform_as).
      * All existing middlewares updated for this breaking change,
future middleware should be written after reading the middlewares docs
again.