grid-contorl: filters: improve terms used for APIs
Filters are set by the filters controllers (filters constructors)
for the current column they are under and not for its specific field.

When we get to the point where we build a query from the column filter,
the affected fields can be any field, and not just the column's one.
Hence, the field terminology, for the column filter state, is
problematic.

clearFieldFilter -> clearColumnFilter
getFieldFilter -> getColumnFilter
setFieldFilter -> setColumnFilter