justdo-chat: when a message sent to a channel with subscribers that set it as unread voluntarily, when another message sent to the channel by another user, mark the channel as iv_unread, to let notifications be triggered
When a channel becomes unread voluntarily, i.e. as a result of the user
himself,
marking it as unread, push notification is now still issued when
another user send
a message to that channel. (When a user mark the channel as unread, he
is usually
aware of what the unread conversation is about, therefore, we need to
help the user
be aware that more things happned in the channel since his last
interaction with it).