grid-data-com: introduce event item-insert-completed
* grid-data-com: the process of inserting a new task (either as a
sibling or as
a child) involves, for performance reasons, additional update following
the
initial insert that sets the actual users of the task.

This update is performed directly to MongoDB (rawCollection) in an async
manner
following the initial insert. Because of that, if a cb is attached to
the Tasks
collection's after insert hook, it will be called before the users of
the task
are set.

Because of that, if the hook needs to know the actual inserted task
users, or,
if the hook performs actions, such as addChild() which behviour is
affected by
the task's users the result will be wrong. In the case of addChild(),
users are
inherited from the parent, hence, if called from the Task after insert
hook,
the created child will have no users other than the task creator.

To overcome that, we now introduced a new event to the grid-data-com
(Server
only) that can be used instead of the after insert hook:

APP.projects._grid_data_com.on("item-insert-completed", (task_id) =>
{}).

That event is called after the raw updates are completed.