grid-data: grid-data-com: fix an issue that caused the raw fields not to add when a single user is being added/removed to a task
Fixes the following issue: When a user is removed from a JustDo, all the
tasks of that JustDo stop belonging to that user. If that user was added
back, if that user still had cached init-payload, when the user opened
the JustDo again, tasks to which that user belonged to, before the
removal from the JustDo, were still shown as if the user still belonged
to them.